Move graph rendering code to its own compilation unit
