+++ src/OpenLoco/CMakeLists.txt
@@ -274,6 +274,7 @@ set(OLOCO_SOURCES
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Title.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Tutorial.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Chart.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Cursor.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Dropdown.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/LastMapWindowAttributes.cpp"
@@ -641,6 +642,7 @@ set(OLOCO_HEADERS
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Tutorial.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Types.hpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui.h"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Chart.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Cursor.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Dropdown.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/LastMapWindowAttributes.h"
